refactor: move menu to separate file, add storybook for it and remove sass
