+++ package.json
@@ -23,7 +23,6 @@
     "bignumber.js": "^9.0.1",
     "mobx": "^6.0.4",
     "mobx-react": "^7.0.5",
-    "node-sass": "4.14.1",
     "qrcode.react": "^1.0.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -45,6 +44,7 @@
     "@storybook/preset-create-react-app": "^3.1.6",
     "@storybook/react": "^6.1.21",
     "@storybook/theming": "^6.1.21",
+    "storybook-react-router": "^1.0.8",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
@@ -55,6 +55,7 @@
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.5",
     "@types/socket.io-client": "^1.4.34",
+    "@types/storybook-react-router": "1.0.0",
     "@typescript-eslint/eslint-plugin": "^4.2.0",
     "@typescript-eslint/parser": "^4.2.0",
     "cross-env": "^7.0.2",
