+++ src/index.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./styles/styles.scss";
 import App from "./App";
 
 ReactDOM.render(
