--- src/styles/abstracts/_mixins.scss
@@ -1,58 +1,0 @@
-@mixin clearfix {
-  &::after {
-    content: "";
-    display: table;
-    clear: both;
-  }
-}
-
-@mixin absCenter {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-@mixin respond($breakpoint) {
-  @if $breakpoint == small-phone {
-    @media only screen and (max-width: 22.8em) {
-      @content;
-    } //max 365px
-  }
-
-  @if $breakpoint == phone {
-    @media only screen and (max-width: 37.5em) {
-      @content;
-    } //max 600px
-  }
-
-  @if $breakpoint == tab-port {
-    @media only screen and (max-width: 56.25em) {
-      @content;
-    } // max 900px
-  }
-
-  @if $breakpoint == tab-land {
-    @media only screen and (max-width: 75em) {
-      @content;
-    } // max 1200px
-  }
-
-  @if $breakpoint == medium-desktop {
-    @media only screen and (max-width: 93.75em) {
-      @content;
-    } // max 1500px
-  }
-
-  @if $breakpoint == desktop {
-    @media only screen and (max-width: 112.5em) {
-      @content;
-    } //max 1800px
-  }
-
-  @if $breakpoint == big-desktop {
-    @media only screen and (min-width: 125em) {
-      @content;
-    } //min width 2000px
-  }
-}
