--- src/styles/abstracts/_variables.scss
@@ -1,5 +1,0 @@
-// Grid
-$grid-width: 114rem;
-$gutter-vertical: 0rem;
-$gutter-vertical-small: 0rem;
-$gutter-horizontal: 0rem;
