--- src/styles/styles.scss
@@ -1,6 +1,0 @@
-@import "./abstracts/mixins";
-@import "./abstracts/variables";
-
-@import "./base/base";
-
-@import "./layout/grid";
